+++ fastly/objectstorage/doc.go
@@ -1,0 +1,3 @@
+// Package objectstorage contains subpackages which offer various operations to
+// configure storage objects.
+package objectstorage