+++ fastly/waf.go
@@ -63,8 +63,8 @@ type ListWAFsInput struct {
 func (i *ListWAFsInput) formatFilters() map[string]string {
 	result := map[string]string{}
 	pairings := map[string]any{
-		"page[size]":                     i.PageSize,
-		"page[number]":                   i.PageNumber,
+		jsonapi.QueryParamPageSize:       i.PageSize,
+		jsonapi.QueryParamPageNumber:     i.PageNumber,
 		"filter[service_id]":             i.FilterService,
 		"filter[service_version_number]": i.FilterVersion,
 		"include":                        i.Include,